smartcityportal/requirements.txt
asgiref==3.11.1
cloudinary==1.44.2
dj-database-url==3.1.2
Django==4.2.27
gunicorn==25.1.0
packaging==26.0
django-cloudinary-storage==0.3.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
